+++ frontend/src/Pages/MyAccount/AllMyReviews.tsx
@@ -45,28 +45,7 @@ const AllMyReviews = () => {
     return result;
   };
 
-  function sortByCreatedDate(a: ReviewGet, b: ReviewGet) {
-    const createDateA = new Date(a.createdAt!).getTime();
-    const createDateB = new Date(b.createdAt!).getTime();
-    // console.log(`Created Date of 1 ${a.film.title} is ${a.createdAt} or ${createDateA}`);
-    // console.log(`Created Date of 2 ${b.film.title} is ${b.createdAt} or ${createDateB}`);
-    return createDateB-createDateA;
-  }
-
-  const sortedReviews = [...reviews].sort((a, b) => {
-    if (a.status === 2 && b.status !== 2) return -1;
-    if (a.status !== 2 && b.status === 2) return 1;
-
-    if (a.startDate === b.startDate) {
-     return sortByCreatedDate(a,b)
-    }
-
-    const dateA = new Date(a.startDate!).getTime();
-    const dateB = new Date(b.startDate!).getTime();
-    return dateB - dateA;
-  });
-
-  const filmIndexes = getFilmIndexes(sortedReviews).reverse();
+  const filmIndexes = getFilmIndexes(reviews).reverse();
 
   const fetchReviews = async () => {
     try {
@@ -80,11 +59,13 @@ const AllMyReviews = () => {
   };
 
   const handleDelete = async (id: number) => {
-    try {
-      await deleteMyReviewAPI(id);
-      fetchReviews();
-    } catch (error) {
-      toast.error("Ошибка при удалении");
+    if (window.confirm("Ты уверен?")) {
+      try {
+        await deleteMyReviewAPI(id);
+        fetchReviews();
+      } catch (error) {
+        toast.error("Ошибка при удалении");
+      }
     }
   };
 
@@ -100,7 +81,7 @@ const AllMyReviews = () => {
         <p className="text-gray-600">Вы ещё не оставили ни одного отзыва.</p>
       ) : (
         <ul className="space-y-6">
-          {sortedReviews.map((review, index) => (
+          {reviews.map((review, index) => (
             <li className="flex items-center gap-3" key={review.id}>
               <ReviewCard
                 index={filmIndexes[index]}
